c: Move checking assertions from recursion when forming composite types to avoid ICE.
The checking assertion in composite_type_internal for structures and unions may
fail if there are self-referential types.  To avoid this, we move them out of
the recursion.  This should also be more efficient and covers other types.
We have to ignore some cases where we form composite types with qualifiers
not matching (PR120510).

gcc/c/ChangeLog:
	* c-typeck.cc (composite_type_internal,composite_type): Move
	checking assertions.

gcc/testsuite/ChangeLog:
	* gcc.dg/gnu23-tag-composite-6.c: Update.