switch-conversion: Mark CSWTCH as mergeable [PR120451]
When we have a smallish CSWTCH, it could be placed in the rodata.cst16
section so it can be merged with other constants across TUs.

The fix is simple; just mark the decl as mergable (DECL_MERGEABLE).
DECL_MERGEABLE was added with r14-1500-g4d935f52b0d5c0 specifically
to improve these kind of decls.

	PR tree-optimization/120451

gcc/ChangeLog:

	* tree-switch-conversion.cc (switch_conversion::build_one_array): Mark
	the newly created decl as mergable.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/cswtch-6.c: New test.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>