Adjust VxWorks special case in testsuite check_weak_available
check_weak_available was reporting weak symbols unsupported
for vxworks unconditionally while they are actually supported
vxworks 7 now (assumed >= r2).  This change adjusts the
predicate to reflect that.

We used to believe we should distinguish kernel and rtp modes,
and experiments showed that this distinction is actually
counterproductive for the testsuite's purposes.

This allows a few extra tests to run (and pass :), in particular
in g++.dg/modules.

2021-02-03  Olivier Hainque  <hainque@adacore.com>

gcc/testsuite/

	* lib/target-supports.exp (check_weak_available):
	Return 1 for VxWorks7.