aarch64: encapsulate note.gnu.property emission into a class
The code emitting the GNU properties was moved to a separate file to
improve modularity and "releave" the 31000-lines long aarch64.cc file
from a few lines.

It introduces a new namespace "aarch64::" for AArch64 backend which
reduce the length of function names by not prepending 'aarch64_' to
each of them.

gcc/ChangeLog:

	* Makefile.in: Add missing declaration of BACKEND_H.
	* config.gcc: Add aarch64-elf-metadata.o to extra_objs.
	* config/aarch64/aarch64-elf-metadata.h: New file
	* config/aarch64/aarch64-elf-metadata.cc: New file.
	* config/aarch64/aarch64.cc
	(GNU_PROPERTY_AARCH64_FEATURE_1_AND): Removed.
	(GNU_PROPERTY_AARCH64_FEATURE_1_BTI): Likewise.
	(GNU_PROPERTY_AARCH64_FEATURE_1_PAC): Likewise.
	(GNU_PROPERTY_AARCH64_FEATURE_1_GCS): Likewise.
	(aarch64_file_end_indicate_exec_stack): Move GNU properties code to
	aarch64-elf-metadata.cc
	* config/aarch64/t-aarch64: Declare target aarch64-elf-metadata.o