[AutoFDO] Profile merging for clone test
This patch introduces a new testcase to verify the merging of profiles
is performed for cloned functions.

Since this is invoked very early, before the pass manager, we need to
set up the dumping explicitly. This is similar to the handling in
finish_optimization_passes.

gcc/ChangeLog:

	* auto-profile.cc (autofdo_source_profile::read): Dump message
	while merging profile.
	* pass_manager.h (get_pass_auto_profile): New.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-prof/clone-merge-1.c: New test.

Signed-off-by: Kugan Vivekanandarajah <kvivekananda@nvidia.com>